move file related listeners to their own space
these listener methods don't have to do with `FileMetadata`, so they should be
reorganized out into their own class. this makes it easier for downstream apps
to override behaviors that are only related to files.
